fix error

+++ src/configuration.cc
@@ -709,7 +709,7 @@ void config::parse_option(std::string const option, std::string const value_stri
     else if (option.substr(5, 7) == "verbose")
       Config::set().energy.dftb.verbosity = std::stoi(value_string);
     else if (option.substr(5, 6) == "scctol")
-      Config::set().energy.dftb.scctol = std::stoi(value_string);
+      Config::set().energy.dftb.scctol = std::stod(value_string);
   }
 
   //Gaussian options
